Resolve 28 conflicts across Razor views, CSS, and shared layouts:
keep design system tokens/components, preserve master's functional
features (dashboard v2, agendamento, auth premium layout, dark mode)

+++ EasyStock.Domain/Exceptions/QuantidadeInvalidaException.cs
@@ -4,7 +4,7 @@ namespace EasyStock.Domain.Exceptions
 {
     public class QuantidadeInvalidaException : RegraDeDominioVioladaException
     {
-        public QuantidadeInvalidaException(int quantidade)
+        public QuantidadeInvalidaException(decimal quantidade)
         : base($"Quantidade inválida: {quantidade}.")
         {
         }
